requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
numpy<2.0
ffmpeg-python
